+++ sw.js
@@ -13,7 +13,7 @@
  */
 'use strict';
 
-var VERSION = 'v62';
+var VERSION = 'v63';
 var RUNTIME_CACHE_VERSION = 'v2';
 var SHELL_CACHE = 'stormscope-shell-' + VERSION;
 var TILE_CACHE = 'stormscope-tiles-' + RUNTIME_CACHE_VERSION;
@@ -44,6 +44,7 @@ var SHELL_ASSETS = [
   './js/context-layers.js',
   './js/tropical-cyclones.js',
   './js/flood-outlooks.js',
+  './js/local-overlays.js',
   './js/diagnostics.js',
   './js/spatial-query.js',
   './js/app.js',